requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0

# Object Detection
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=9.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
tensorboard>=2.13.0
albumentations>=1.3.0
